+++ src/main.ts
@@ -1,4 +1,0 @@
-import { init } from './app';
-import './style.scss';
-
-init();